modify logics of generateEncryptKey and getEncryptionKey
